Renames console util to swri_console util. Refs #231.
