+++ swri_console_util/include/swri_console_util/progress_bar.h
@@ -35,7 +35,7 @@
 
 #include <ros/ros.h>
 
-namespace console_util
+namespace swri_console_util
 {
   class ProgressBar
   {
