+++ swri_console_util/package.xml
@@ -1,9 +1,9 @@
 <package format="2">
-  <name>console_util</name>
+  <name>swri_console_util</name>
   <version>0.0.1</version>
   <description>
 
-     console_util
+     swri_console_util
 
   </description>
   <author>Marc Alban</author>
